requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
PyJWT==2.8.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
statsmodels==0.14.1
lifelines==0.28.0
shap==0.44.1
openpyxl==3.1.2
python-dotenv==1.0.0
